+++ src/module/settings.ts
@@ -532,7 +532,7 @@ export let fetchParams = () => {
   if (configSettings.midiUnconsciousCondition === undefined) configSettings.midiUnconsciousCondition = "none";
   // Fix for typo in en.json
   if (configSettings.autoTarget === "wallsBlockIgnoreIncapcitated") configSettings.autoTarget = "wallsBlockIgnoreIncapacitated";
-  if (configSettings.autoTarget === "wallsBlockIgnoreIncapacitated") configSettings.autoTarget = "alwaysIgnoreIncapacitated";
+  if (configSettings.autoTarget === "wallsBlockIgnoreIncapacitated") configSettings.autoTarget = "wallsBlockIgnoreIncapacitated";
   if (configSettings.autoTarget === "alwaysIgnoreIncapcitated") configSettings.autoTarget = "alwaysIgnoreIncapacitated";
   if (configSettings.midiFieldsTab === undefined) configSettings.midiFieldsTab = true;
   configSettings.v3DamageApplication = true;
@@ -559,8 +559,6 @@ export let fetchParams = () => {
   targetConfirmation = game.settings.get("midi-qol", "TargetConfirmation");
   if (configSettings.griddedGridless === undefined) configSettings.griddedGridless = false;
   if (configSettings.gridlessFudge === undefined) configSettings.gridlessFudge = 0;
-  //@ts-expect-error
-  if (game.release.generation >= 12) configSettings.gridlessFudge = undefined;
   if (configSettings.concentrationIncapacitatedConditionCheck === undefined) configSettings.concentrationIncapacitatedConditionCheck = false;
   if (targetConfirmation === undefined || typeof targetConfirmation === "string" || targetConfirmation instanceof String) targetConfirmation = {
     enabled: false,